Isolation of web context, specially for admin components. Web context is child spring context of parent that is represented by OpenHubApplication configuration. Separation of context also reflects security configuration.
Changes:
- added URL constants
- version printer improvement
- camel config - move servlet registration
- web service config - move servlet registration
- OpenHubApplication - create parent/child context hierarchy
	- WebContextConfig as child context configurer, it handles org.openhubframework.openhub.web and org.openhubframework.openhub.admin packages
		- registers beans, servlets and filters
	- WebConfigurer - is responsible for configuration of web runtime layer of OpenHub, for example servlet registration of camel connectors.
- separation of security configuration - authentication manager GlobalSecurityConfig and for root (WebSecurityConfig) and admin context (AdminSecurityConfig)
	- GlobalSecurityConfig - auto-configuration of authentication manager, it is simple possible to override this with custom implementation
	- WebSecurityConfig - accessDecisionManager and authorizationPolicy are exposed only when the specified bean classes and/or names are
  		not already contained in the BeanFactory.
	- AdminSecurityConfig - configuration of security for /web/admin paths
- new package hierarchy definition
	- root context - main configuration of OpenHub - org.openhubframework.openhub.config
	- child context
		- web  - org.openhubframework.openhub.web auto-scan
	    - admin - org.openhubframework.openhub.admin auto-scan
- application.properties
	- identity of application
	- configure live beans view for /beans actuator to reflect current parent-child hierarchy
	- management path is /mgmt
	- white label error page disable
	- deactivate X-Application-Context HTTP Header
- freemarker templates - corrected login and logout handling

Issue: OHFJIRA-40
