+++ core/src/main/java/org/openhubframework/openhub/core/common/version/VersionPrinter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014 the original author or authors.
+ * Copyright 2014-2017 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -52,6 +52,6 @@ public class VersionPrinter {
             versionStr = versions[0].getFullVersion() + " (" + versions[0].getDate() + ")";
         }
 
-        LOG.info("OpenHub version: " + versionStr);
+        LOG.info("\n\nOpenHub version: " + versionStr + "\n");
     }
 }
