+++ web-admin/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,3 @@
+# Auto Configure
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+	org.openhubframework.openhub.config.GlobalSecurityConfig
