+++ web-admin/src/main/resources/application.properties
@@ -36,6 +36,17 @@ logging.level.org.openhubframework.openhub.admin.web.filter.RequestResponseLoggi
 
 
 # ===============================
+# = APPLICATION
+# ===============================
+# IDENTITY (ContextIdApplicationContextInitializer)
+#spring.application.index= # Application index.
+spring.application.name=${info.app.name}
+# live beans view exposure, supported by /beans actuator
+# see http://docs.spring.io/spring/docs/current/javadoc-api/org/springframework/context/support/LiveBeansView.html
+spring.liveBeansView.mbeanDomain=hierarchy
+
+
+# ===============================
 # = JPA / HIBERNATE
 # ===============================
 
@@ -127,6 +138,24 @@ endpoints.jolokia.sensitive=true
 # JOLOKIA (JolokiaProperties)
 jolokia.config.debug=true
 #jolokia.config.*= # See Jolokia manual
+
+
+# ===============================
+# = MANAGEMENT HTTP SERVER
+# ===============================
+# MANAGEMENT HTTP SERVER (ManagementServerProperties)
+# Add the "X-Application-Context" HTTP header in each response.
+management.add-application-context-header=false
+# Management endpoint context-path. For instance `/actuator`
+management.context-path=/mgmt
+
+
+# ===============================
+# = SERVER CONFIGURATION
+# ===============================
+# EMBEDDED SERVER CONFIGURATION (ServerProperties)
+# Enable the default error page displayed in browsers in case of a server error.
+server.error.whitelabel.enabled=false
 
 
 # ===============================
@@ -222,7 +251,7 @@ camel.springboot.name=camelContext
 
 
 # ===============================
-# = throttling
+# = THROTTLING
 # ===============================
 
 ###############################################################################
@@ -251,10 +280,10 @@ throttling.sourceSystem.*=300/60
 throttling.*.syncHello=15/60
 throttling.*.asyncHello=50/60
 
-
+        
 
 # ===============================
-# = alerts
+# = ALERTS
 # ===============================
 
 ###############################################################################
